Refactor UserCard APIFields models
