+++ src/client/lib/authAPI/index.ts
@@ -1,6 +1,7 @@
 import { getAPIWithAuth } from '../fetchAPI'
 
-import { SHARED_APIFields__GetUserCard, SHARED_APIFields__LifetimeStats } from '@/shared/models/apiFields'
+import { SHARED_APIFields__LifetimeStats } from '@/shared/models/apiFields'
+import { SHARED_APIFields__GetUserCard } from '@/shared/models/apiFields/userCard'
 
 /*
 Authenticated APIs (must provide access token)
