+++ src/shared/models/apiFields/userCard.ts
@@ -1,0 +1,7 @@
+import { SHARED_Model__UserCard } from '../models/UserCard'
+
+import { SHARED_APIFields__BASE } from '.'
+
+export interface SHARED_APIFields__GetUserCard extends SHARED_APIFields__BASE {
+    userCard?: SHARED_Model__UserCard
+}
